Add automatic version bumping script
Makefile default rule now calls `autobump` and `index` targets. To
use this, call `make` after changing any of the colour schemes.

The way it works is versions.log contains the md5 sums of the
color scheme files. The autobump script checks the old sum from
versions.log against the current color scheme file and if any of
them changed, it opens the color scheme and increments the version
number in it and then stores the newly computed sum (after the bump)
back into versions.log. When the `index` rule is run in the Makefile,
it updates the `index.json` and `index.json.md5` files based on the
changes.